requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
bcrypt==4.2.0
blinker==1.8.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
distro==1.9.0
Flask==3.0.3
Flask-Cors==5.0.0
fonttools==4.47.0
frozenlist==1.4.1
google-api-core==2.19.2
google-auth==2.34.0
google-cloud-vision==3.7.4
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
kiwisolver==1.4.5
MarkupSafe==2.1.5
matplotlib==3.8.2
multidict==6.1.0
numpy==1.26.2
openai==0.28.0
packaging==23.2
pandas==2.1.4
Pillow==10.1.0
proto-plus==1.24.0
protobuf==5.28.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.9.1
pydantic_core==2.23.3
pyparsing==3.1.1
pyquaternion==0.9.9
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2023.3.post1
PyJWT==2.8.0
cryptography==41.0.0
requests==2.32.3
rsa==4.9
six==1.16.0
sniffio==1.3.1
tqdm==4.66.5
typing_extensions==4.12.2
tzdata==2023.4
urllib3==2.2.2
Werkzeug==3.0.4
yarl==1.11.1
alembic==1.13.3
